Fixed Lua LSP issue; refactored NvimLspConfig into cmp.lua( which is now named CompletionAndLSP.lua )
Messing around with Nui.lua; Creating a Terminal Popup window. Where
instead of 'pushing' the main neovim window up. This will instead hover
over the main window. So far, I'm able to create the Nui popup window,
create a hidden scratch buffer, attach a terminal to it, and push the
buffer into the new Nui window.

Next, I need to refine the Key command controls for this window, and
figure out of to make the window hide when Neovim refocuses back on the
main window