At the moment, I was able to get the border color to change depending on which mode you were in ( normal vs terminal/insert ) But, I found a glaring flaw..
The Terminal Window is attached to the window that it was opened in. And
would be confined to the width of said window. And even worse. If you
tried to toggle the terminal on again, if you changed focus to a
differnet window. A 'new' terminal window would open up. But if you went
into 'insert/terminal' mode, the focus would move to the first terminal
window....

Now, I'm seeing if I can just get rid of Nui and just use Neovims built
in API for creating my own Floating Window